bert4rec_service/mlops/requirements.txt
python-dotenv
pandas
numpy
torch
pytorch_lightning
tensorboard
tensorboardX
fastapi
celery
redis
uvicorn
gdown